+++ src/ccc/data_refinement.h
@@ -3,11 +3,21 @@
 
 #pragma once
 
+#include <functional>
+
 #include "elf.h"
 #include "symbol_database.h"
 
 namespace ccc {
 
-void refine_variables(SymbolDatabase& database, const std::vector<ElfFile*>& elves);
+struct RefinedData {
+	std::string field_name;
+	std::variant<std::string, std::vector<RefinedData>> value;
+};
+
+using ReadVirtualFunc = std::function<Result<void>(u8* dest, u32 address, u32 size)>;
+
+bool can_refine_variable(const Variable& variable);
+Result<RefinedData> refine_variable(const Variable& variable, const SymbolDatabase& database, const ReadVirtualFunc& read_virtual);
 
 }