+++ src/ccc/dependency.cpp
@@ -95,15 +95,21 @@ static void map_types_to_files_based_on_reference_count_single_pass(SymbolDataba
 			} else {
 				for(const Function& function : database.functions.span(file->functions())) {
 					if(function.storage_class != ast::SC_STATIC) {
-						ast::for_each_node(function.type(), ast::PREORDER_TRAVERSAL, count_references);
+						if(function.type_ptr()) {
+							ast::for_each_node(function.type(), ast::PREORDER_TRAVERSAL, count_references);
+						}
 						for(const ParameterVariable& parameter_variable : database.parameter_variables.span(function.parameter_variables())) {
-							ast::for_each_node(parameter_variable.type(), ast::PREORDER_TRAVERSAL, count_references);
+							if(parameter_variable.type_ptr()) {
+								ast::for_each_node(parameter_variable.type(), ast::PREORDER_TRAVERSAL, count_references);
+							}
 						}
 					}
 				}
 				for(const GlobalVariable& global_variable : database.global_variables.span(file->globals_variables())) {
 					if(global_variable.storage_class != ast::SC_STATIC) {
-						ast::for_each_node(global_variable.type(), ast::PREORDER_TRAVERSAL, count_references);
+						if(global_variable.type_ptr()) {
+							ast::for_each_node(global_variable.type(), ast::PREORDER_TRAVERSAL, count_references);
+						}
 					}
 				}
 			}